+++ SensorNode/config.h
@@ -8,7 +8,7 @@ void cfg_init();
 uchar cfg_node_id();
 void cfg_set_node_id(uchar val);
 
-ulong cfg_ds1820_period();
-void cfg_set_ds1820_period(ulong val);
+ushort cfg_ds1820_period();
+void cfg_set_ds1820_period(ushort val);
 
 #endif /* _CONFIG_H_ */
